testing: reorg; unify access to Passphrase
